Consolidate keyboard shortcuts and add preventDefault to avoid browser conflicts
- Move duplicate KeyboardShortcuts components to shared location at src/editor/components/keyboard-shortcuts.tsx
- Remove three identical keyboard-shortcuts.tsx files from individual node properties panels
- Add preventDefault() to avoid conflicts with browser shortcuts for Cmd+Enter/Ctrl+Enter
- Maintain cross-platform compatibility (Cmd+Enter on Mac, Ctrl+Enter on others)
- Preserve component-based approach for better explicitness and debugging

Fixes #995: Consider preventing default behavior and adding event scope validation
