+++ internal-packages/workflow-designer-ui/src/editor/components/keyboard-shortcuts.tsx
@@ -12,6 +12,7 @@ export function KeyboardShortcuts({
 				event.key === "Enter" &&
 				((isMac && event.metaKey) || (!isMac && event.ctrlKey))
 			) {
+				event.preventDefault();
 				generate();
 			}
 		};
